backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Fillsogood <jdd05273@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
django = "^5.0.3"
pyjwt = "^2.8.0"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
mysqlclient = "^2.2.4"
django-cors-headers = "^4.3.1"
rest-framework-simplejwt = "^0.0.2"
drf-spectacular = "^0.27.1"
openai = "^1.16.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
